Unroll the *.tomcat packages - Pull implementations up to their interfaces and migrate them to classes - Simplify code thanks to lambda runnables - Fix Spring bean definitions
