+++ src/main/java/org/opendatakit/aggregate/client/form/KmlSelection.java
@@ -1,6 +1,12 @@
 package org.opendatakit.aggregate.client.form;
 
+import static org.opendatakit.aggregate.client.form.KmlSelection.KmlOptionType.GEO_POINT;
+import static org.opendatakit.aggregate.client.form.KmlSelection.KmlOptionType.GEO_TRACE_N_SHAPE;
+import static org.opendatakit.aggregate.client.form.KmlSelection.KmlOptionType.UNDEFINED;
+import static org.opendatakit.aggregate.client.form.KmlSelection.KmlOptionType.valueOf;
+
 import java.io.Serializable;
+import java.util.Optional;
 
 public class KmlSelection implements Serializable {
   /**
@@ -29,6 +35,39 @@ public class KmlSelection implements Serializable {
     clear();
   }
 
+  public static Optional<KmlSelection> from(String encodedString) {
+    String[] fieldStrings = encodedString.split(DELIMITER);
+
+    if (fieldStrings.length < 3)
+      return Optional.empty();
+
+    String formId = fieldStrings[0];
+    KmlOptionType type = valueOf(fieldStrings[1]);
+    String geoElementKey = fieldStrings[2];
+
+    // check basic constraints on size and nullability
+    if (formId == null || geoElementKey == null)
+      return Optional.empty();
+
+    if (type == GEO_POINT && fieldStrings.length != 5)
+      return Optional.empty();
+
+    if (type == GEO_TRACE_N_SHAPE && fieldStrings.length != 4)
+      return Optional.empty();
+
+    if (type != GEO_POINT && type != GEO_TRACE_N_SHAPE)
+      return Optional.empty();
+
+    KmlSelection selection = new KmlSelection(formId);
+    if (type == GEO_POINT)
+      selection.setGeoPointSelections(geoElementKey, fieldStrings[3], fieldStrings[4]);
+
+    if (type == GEO_TRACE_N_SHAPE)
+      selection.setGeoTraceNShapeSelections(geoElementKey, fieldStrings[3]);
+
+    return Optional.of(selection);
+  }
+
   public static KmlSelection createKmlSelectionFromEncodedString(String encodedString) {
     String[] fieldStrings = encodedString.split(DELIMITER);
 
@@ -37,21 +76,21 @@ public class KmlSelection implements Serializable {
     }
 
     String formId = fieldStrings[0];
-    KmlOptionType type = KmlOptionType.valueOf(fieldStrings[1]);
+    KmlOptionType type = valueOf(fieldStrings[1]);
     String geoElementKey = fieldStrings[2];
 
     // check basic constraints on size and nullability
     if ((type == null) || (formId == null) || (geoElementKey == null) ||
-        (type == KmlOptionType.GEO_POINT && fieldStrings.length != 5) ||
-        (type == KmlOptionType.GEO_TRACE_N_SHAPE && fieldStrings.length != 4)) {
+        (type == GEO_POINT && fieldStrings.length != 5) ||
+        (type == GEO_TRACE_N_SHAPE && fieldStrings.length != 4)) {
       return null;
     }
 
 
     KmlSelection selection = new KmlSelection(formId);
-    if (type == KmlOptionType.GEO_POINT) {
+    if (type == GEO_POINT) {
       selection.setGeoPointSelections(geoElementKey, fieldStrings[3], fieldStrings[4]);
-    } else if (type == KmlOptionType.GEO_TRACE_N_SHAPE) {
+    } else if (type == GEO_TRACE_N_SHAPE) {
       selection.setGeoTraceNShapeSelections(geoElementKey, fieldStrings[3]);
     } else {
       return null;
@@ -61,7 +100,7 @@ public class KmlSelection implements Serializable {
   }
 
   private void clear() {
-    geoType = KmlOptionType.UNDEFINED;
+    geoType = UNDEFINED;
     geoElementKey = null;
     titleElementKey = null;
     binaryElementKey = null;
@@ -71,7 +110,7 @@ public class KmlSelection implements Serializable {
   public void setGeoPointSelections(String geoKey, String titleKey,
                                     String binaryKey) {
     clear();
-    geoType = KmlOptionType.GEO_POINT;
+    geoType = GEO_POINT;
     this.geoElementKey = geoKey;
     this.titleElementKey = titleKey;
     this.binaryElementKey = binaryKey;
@@ -79,7 +118,7 @@ public class KmlSelection implements Serializable {
 
   public void setGeoTraceNShapeSelections(String geoKey, String nameKey) {
     clear();
-    geoType = KmlOptionType.GEO_TRACE_N_SHAPE;
+    geoType = GEO_TRACE_N_SHAPE;
     this.geoElementKey = geoKey;
     this.nameElementKey = nameKey;
   }
