+++ src/main/java/org/opendatakit/aggregate/task/KmlGenerator.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2010 University of Washington
+ * Copyright (C) 2018 Nafundi
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -15,22 +16,38 @@
  */
 package org.opendatakit.aggregate.task;
 
+import static java.util.Collections.emptyList;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import org.opendatakit.aggregate.client.form.KmlSelection;
+import org.opendatakit.aggregate.constants.BeanDefs;
 import org.opendatakit.aggregate.form.IForm;
 import org.opendatakit.aggregate.form.PersistentResults;
 import org.opendatakit.common.persistence.exception.ODKDatastoreException;
 import org.opendatakit.common.web.CallingContext;
 
-/**
- * API for creating and restarting Kml generation tasks.
- *
- * @author wbrunette@gmail.com
- * @author mitchellsundt@gmail.com
- */
-public interface KmlGenerator {
+public class KmlGenerator {
 
-  public static final String KML_SELECTIONS_KEY = "KMLSELECTIONS";
-  public static final String KML_SELECTIONS_DELIMITER = "###";
+  public void createKmlTask(IForm form, PersistentResults persistentResults, long attemptCount, CallingContext cc) throws ODKDatastoreException {
+    Watchdog wd = (Watchdog) cc.getBean(BeanDefs.WATCHDOG);
+    KmlWorkerImpl worker = new KmlWorkerImpl(form, persistentResults.getSubmissionKey(), attemptCount, getKmlElementsToInclude(persistentResults), wd.getCallingContext());
+    AggregrateThreadExecutor.getAggregateThreadExecutor().execute(worker::generateKml);
+  }
 
-  public void createKmlTask(IForm form, PersistentResults persistentResult,
-                            long attemptCount, CallingContext cc) throws ODKDatastoreException;
+  private static List<KmlSelection> getKmlElementsToInclude(PersistentResults persistentResults) throws ODKDatastoreException {
+    Map<String, String> params = persistentResults.getRequestParameters();
+    if (params.get("KMLSELECTIONS") == null)
+      return emptyList();
+
+    String[] kmlSelection = params.get("KMLSELECTIONS").split("###");
+    return Stream.of(kmlSelection)
+        .map(KmlSelection::from)
+        .filter(Optional::isPresent)
+        .map(Optional::get)
+        .collect(Collectors.toList());
+  }
 }
