+++ src/main/java/org/opendatakit/aggregate/task/PurgeOlderSubmissions.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2011 University of Washington.
+ * Copyright (C) 2018 Nafundi
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -15,19 +16,18 @@
  */
 package org.opendatakit.aggregate.task;
 
+import org.opendatakit.aggregate.constants.BeanDefs;
 import org.opendatakit.aggregate.form.IForm;
 import org.opendatakit.aggregate.submission.SubmissionKey;
 import org.opendatakit.common.web.CallingContext;
 
-/**
- * Create the task that will purge all of a form's submissions older than the given date.
- *
- * @author mitchellsundt@gmail.com
- */
-public interface PurgeOlderSubmissions {
+public class PurgeOlderSubmissions {
 
   public static final String PURGE_DATE = "purgeBefore";
 
-  public void createPurgeOlderSubmissionsTask(IForm form, SubmissionKey miscTasksKey,
-                                              long attemptCount, CallingContext cc);
+  public final void createPurgeOlderSubmissionsTask(IForm form, SubmissionKey miscTasksKey, long attemptCount, CallingContext cc) {
+    Watchdog wd = (Watchdog) cc.getBean(BeanDefs.WATCHDOG);
+    PurgeOlderSubmissionsWorkerImpl worker = new PurgeOlderSubmissionsWorkerImpl(form, miscTasksKey, attemptCount, wd.getCallingContext());
+    AggregrateThreadExecutor.getAggregateThreadExecutor().execute(worker::purgeOlderSubmissions);
+  }
 }
