--- src/main/java/org/opendatakit/aggregate/util/tomcat/ImageUtilImpl.java
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) 2011 University of Washington
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package org.opendatakit.aggregate.util.tomcat;
-
-import org.opendatakit.aggregate.util.ImageUtil;
-import org.opendatakit.tomcatutil.ImageResize;
-
-public class ImageUtilImpl implements ImageUtil {
-
-  @Override
-  public byte[] resizeImage(byte[] imageBlob, int width, int height) {
-    ImageResize resize = new ImageResize();
-    return resize.resizeImage(imageBlob, width, height);
-  }
-
-}
